Refactor: extract test config helpers into config_helpers.py
Move findAvailableArchs, findConfigs, configMarks, walkDict,
markNamed, and get_rocm_version_or_none from test_config.py into a
dedicated config_helpers module. No behavioral changes.
